# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Persistent shared memory infrastructure for AI agents. Stores observations in PostgreSQL + pgvector, exposes MCP tools via HTTP/SSE.",
        "title": "Engram API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:37777",
    "basePath": "/",
    "paths": {
...
        "/api/graph/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns graph statistics for the dashboard, using relation data to compute knowledge graph metrics.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Graph"
                ],
                "summary": "Get graph statistics",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/models": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns available embedding models with default and current model info. Cacheable for 1 hour.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Observations"
                ],
                "summary": "List embedding models",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/observations": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns recent observations with optional semantic search via vector store. Supports pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Observations"
                ],
                "summary": "List observations",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by project",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Semantic search query",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results (default 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Pagination offset",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/projects": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns all known projects. Response is cacheable for 5 minutes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Observations"
                ],
                "summary": "List projects",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "object"
                            }
                        }
                    },
                    "500": {
                        "description": "internal error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
        "/api/search/analytics": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns comprehensive search analytics including vector search rate, zero result rate, top keywords, and query type distribution.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get search analytics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by project",
                        "name": "project",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/api/search/recent": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns recent search queries for analytics purposes.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get recent search queries",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by project",
                        "name": "project",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Number of results (default 20)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/stats": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns comprehensive worker statistics including uptime, memory, database health, vector cache, graph, and rate limiter stats.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get worker statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter stats by project",
                        "name": "project",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "bad request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/stats/retrieval": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns detailed retrieval statistics including hit rates and latency.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Analytics"
                ],
                "summary": "Get retrieval statistics",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by project",
                        "name": "project",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/api/types": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Returns the canonical list of observation and concept types. Cacheable for 24 hours.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Observations"
                ],
                "summary": "List observation and concept types",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
...
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "X-Auth-Token",
            "in": "header"
        }
    }
}